+++ doorstop/cli/commands.py
@@ -8,6 +8,7 @@ from doorstop.cli import utilities
 from doorstop.cli.utilities import show, ask
 from doorstop.core.builder import build
 from doorstop.core import editor, importer, exporter, publisher
+from doorstop import server
 
 log = common.logger(__name__)
 
@@ -22,19 +23,21 @@ def get(name):
         return run
 
 
-def run(args, cwd, err, catch=True):  # pylint: disable=W0613
+def run(args, cwd, error, catch=True):  # pylint: disable=W0613
     """Process arguments and run the `doorstop` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        show("validating tree...", flush=True)
-        tree = build(cwd, root=args.project)
+        show("building tree...", flush=True)
+        tree = build(cwd=cwd, root=args.project)
+        show("loading documents...", flush=True)
         tree.load()
+        show("validating items...", flush=True)
         valid = tree.validate()
     if not success:
         return False
@@ -50,12 +53,12 @@ def run_create(args, cwd, _, catch=True):
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         document = tree.create_document(args.path, args.prefix,
                                         parent=args.parent, digits=args.digits)
     if not success:
@@ -70,12 +73,12 @@ def run_delete(args, cwd, _, catch=True):
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         document = tree.find_document(args.prefix)
         prefix, relpath = document.prefix, document.relpath
         document.delete()
@@ -92,14 +95,27 @@ def run_add(args, cwd, _, catch=True):
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
+        document = tree.find_document(args.prefix)
+        if args.force:
+            log.warn("creating items without the server...")
+        else:
+            server.check()
         for _ in range(args.count):
-            item = tree.add_item(args.prefix, level=args.level)
+            number = 0
+            while number is not None and number < document.next:
+                if number:
+                    log.warn("server is behind, requesting next number...")
+                if args.force:
+                    number = None
+                else:
+                    number = server.get_next_number(args.prefix)
+            item = document.add_item(number=number, level=args.level)
             show("added item: {} ({})".format(item.uid, item.relpath))
 
     if not success:
@@ -113,12 +129,12 @@ def run_remove(args, cwd, _, catch=True):
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         item = tree.find_item(args.uid)
         item.delete()
     if not success:
@@ -129,20 +145,20 @@ def run_remove(args, cwd, _, catch=True):
     return True
 
 
-def run_edit(args, cwd, err, catch=True):
+def run_edit(args, cwd, error, catch=True):
     """Process arguments and run the `doorstop edit` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     item = document = None
-    ext = utilities.get_ext(args, '.yml', '.yml', whole_tree=False, err=err)
+    ext = utilities.get_ext(args, '.yml', '.yml', whole_tree=False, error=error)
 
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         # find item or document
         if not args.document:
             try:
@@ -156,7 +172,7 @@ def run_edit(args, cwd, err, catch=True):
         if item:
             item.edit(tool=args.tool)
         else:
-            _export_import(args, cwd, err, document, ext)
+            _export_import(args, cwd, error, document, ext)
     if not success:
         return False
 
@@ -166,19 +182,19 @@ def run_edit(args, cwd, err, catch=True):
     return True
 
 
-def run_reorder(args, cwd, err, catch=True, _tree=None):
+def run_reorder(args, cwd, error, catch=True, _tree=None):
     """Process arguments and run the `doorstop reorder` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     reordered = False
 
     with utilities.capture(catch=catch) as success:
-        tree = _tree or build(cwd, root=args.project)
+        tree = _tree or build(cwd=cwd, root=args.project)
         document = tree.find_document(args.prefix)
     if not success:
         return False
@@ -203,7 +219,7 @@ def run_reorder(args, cwd, err, catch=True, _tree=None):
             document.index = True  # create index
             relpath = os.path.relpath(document.index, cwd)
             editor.edit(relpath, tool=args.tool)
-            get('reorder')(args, cwd, err, catch=False, _tree=tree)
+            get('reorder')(args, cwd, error, catch=False, _tree=tree)
     if not success:
         show("after fixing the error: doorstop reorder {}".format(document))
         return False
@@ -219,12 +235,12 @@ def run_link(args, cwd, _, catch=True):
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         child, parent = tree.link_items(args.child, args.parent)
     if not success:
         return False
@@ -240,12 +256,12 @@ def run_unlink(args, cwd, _, catch=True):
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         child, parent = tree.unlink_items(args.child, args.parent)
     if not success:
         return False
@@ -256,17 +272,17 @@ def run_unlink(args, cwd, _, catch=True):
     return True
 
 
-def run_clear(args, cwd, err, catch=True):
+def run_clear(args, cwd, error, catch=True):
     """Process arguments and run the `doorstop clear` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        for item in _iter_items(args, cwd, err):
+        for item in _iter_items(args, cwd, error):
             show("clearing item {}'s suspect links...".format(item.uid))
             item.clear()
     if not success:
@@ -275,17 +291,17 @@ def run_clear(args, cwd, err, catch=True):
     return True
 
 
-def run_review(args, cwd, err, catch=True):
+def run_review(args, cwd, error, catch=True):
     """Process arguments and run the `doorstop review` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     with utilities.capture(catch=catch) as success:
-        for item in _iter_items(args, cwd, err):
+        for item in _iter_items(args, cwd, error):
             show("marking item {} as reviewed...".format(item.uid))
             item.review()
     if not success:
@@ -294,35 +310,35 @@ def run_review(args, cwd, err, catch=True):
     return True
 
 
-def run_import(args, cwd, err, catch=True, _tree=None):
+def run_import(args, cwd, error, catch=True, _tree=None):
     """Process arguments and run the `doorstop import` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     document = item = None
 
     # Parse arguments
-    attrs = utilities.literal_eval(args.attrs, err)
-    mapping = utilities.literal_eval(args.map, err)
+    attrs = utilities.literal_eval(args.attrs, error)
+    mapping = utilities.literal_eval(args.map, error)
     if args.path:
         if not args.prefix:
-            err("when [path] specified, [prefix] is also required")
+            error("when [path] specified, [prefix] is also required")
         elif args.document:
-            err("'--document' cannot be used with [path] [prefix]")
+            error("'--document' cannot be used with [path] [prefix]")
         elif args.item:
-            err("'--item' cannot be used with [path] [prefix]")
-        ext = utilities.get_ext(args, None, None, False, err)
+            error("'--item' cannot be used with [path] [prefix]")
+        ext = utilities.get_ext(args, None, None, False, error)
     elif not (args.document or args.item):
-        err("specify [path], '--document', or '--item' to import")
+        error("specify [path], '--document', or '--item' to import")
 
     # Import document or item
     with utilities.capture(catch=catch) as success:
         if args.path:
-            tree = _tree or build(cwd, root=args.project)
+            tree = _tree or build(cwd=cwd, root=args.project)
             document = tree.find_document(args.prefix)
             msg = "importing '{}' into document {}...".format(args.path,
                                                               document)
@@ -349,23 +365,23 @@ def run_import(args, cwd, err, catch=True, _tree=None):
     return True
 
 
-def run_export(args, cwd, err, catch=True):
+def run_export(args, cwd, error, catch=True):
     """Process arguments and run the `doorstop export` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     # Parse arguments
     whole_tree = args.prefix == 'all'
-    ext = utilities.get_ext(args, '.yml', '.csv', whole_tree, err)
+    ext = utilities.get_ext(args, '.yml', '.csv', whole_tree, error)
 
     # Export documents
     with utilities.capture(catch=catch) as success:
         exporter.check(ext)
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         if not whole_tree:
             document = tree.find_document(args.prefix)
     if not success:
@@ -387,30 +403,30 @@ def run_export(args, cwd, err, catch=True):
     # Display to standard output
     else:
         if whole_tree:
-            err("only single documents can be displayed")
+            error("only single documents can be displayed")
         for line in exporter.export_lines(document, ext):
             show(line)
 
     return True
 
 
-def run_publish(args, cwd, err, catch=True):
+def run_publish(args, cwd, error, catch=True):
     """Process arguments and run the `doorstop publish` subcommand.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param catch: catch and log :class:`~doorstop.common.DoorstopError`
 
     """
     # Parse arguments
     whole_tree = args.prefix == 'all'
-    ext = utilities.get_ext(args, '.txt', '.html', whole_tree, err)
+    ext = utilities.get_ext(args, '.txt', '.html', whole_tree, error)
 
     # Publish documents
     with utilities.capture(catch=catch) as success:
         publisher.check(ext)
-        tree = build(cwd, root=args.project)
+        tree = build(cwd=cwd, root=args.project)
         if not whole_tree:
             document = tree.find_document(args.prefix)
     if not success:
@@ -437,19 +453,19 @@ def run_publish(args, cwd, err, catch=True):
     # Display to standard output
     else:
         if whole_tree:
-            err("only single documents can be displayed")
+            error("only single documents can be displayed")
         for line in publisher.publish_lines(document, ext, **kwargs):
             show(line)
 
     return True
 
 
-def _iter_items(args, cwd, err):
+def _iter_items(args, cwd, error):
     """Build a tree and iterate through items.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
 
     Items are filtered to:
 
@@ -466,14 +482,14 @@ def _iter_items(args, cwd, err):
     # Parse arguments
     if args.label == 'all':
         if args.item:
-            err("argument -i/--item: not allowed with 'all'")
+            error("argument -i/--item: not allowed with 'all'")
         if args.document:
-            err("argument -d/--document: not allowed with 'all'")
+            error("argument -d/--document: not allowed with 'all'")
 
     # Build tree
     item = None
     document = None
-    tree = build(cwd, root=args.project)
+    tree = build(cwd=cwd, root=args.project)
 
     # Determine if tree, document, or item was requested
     if args.label != 'all':
@@ -498,12 +514,12 @@ def _iter_items(args, cwd, err):
                 yield item
 
 
-def _export_import(args, cwd, err, document, ext):
+def _export_import(args, cwd, error, document, ext):
     """Edit a document by calling export followed by import.
 
     :param args: Namespace of CLI arguments
     :param cwd: current working directory
-    :param err: function to call for CLI errors
+    :param error: function to call for CLI errors
     :param document: :class:`~doorstop.core.document.Document` to edit
     :param ext: extension for export format
 
@@ -512,7 +528,7 @@ def _export_import(args, cwd, err, document, ext):
     args.prefix = document.prefix
     path = "{}-{}{}".format(args.prefix, int(time.time()), ext)
     args.path = path
-    get('export')(args, cwd, err, catch=False)
+    get('export')(args, cwd, error, catch=False)
 
     # Open the exported file
     editor.edit(path, tool=args.tool)
@@ -521,7 +537,7 @@ def _export_import(args, cwd, err, document, ext):
     if ask("import from '{}'?".format(path)):
         args.attrs = {}
         args.map = {}
-        get('import')(args, cwd, err, catch=False, _tree=document.tree)
+        get('import')(args, cwd, error, catch=False, _tree=document.tree)
         common.delete(path)
     else:
         show("import canceled")