+++ doorstop/core/vcs/veracity.py
@@ -14,11 +14,11 @@ class WorkingCopy(BaseWorkingCopy):  # pragma: no cover (integration test)
     IGNORES = ('.sgignores', '.vvignores')
 
     def lock(self, path):
-        self.call('vv', 'pull')
-        self.call('vv', 'update')
         # TODO: track: http://veracity-scm.com/qa/questions/2034
         msg = "veracity does not support scripted locking: {}".format(path)
-        log.warning(msg)
+        log.info(msg)
+        self.call('vv', 'pull')
+        self.call('vv', 'update')
 
     def save(self, message=None):
         message = message or input("Commit message: ")  # pylint: disable=W0141